requirements.txt
PyMuPDF
pymongo
langchain-community
langchain-openai
streamlit
python-dotenv